review: add ProductListComponent
